==> 04a_calc_prop_feed_per_crop_country_FAO_FB_and_CB_2013.Rmd <==
---
title: "Calculating proportion crop used as feed using 2013 FAO food balance and commodity balances data"
author: "Haley Epperly"
date: "2022-10-26"
output: html_document
editor_options: 
  chunk_output_type: console
---

Objective: Use 2013 data from the FAO Food Balances and Commodity Balances data to determine the proportion of crops being used as feed for which most of the crop is going to the processing category.

Crops we need to calculate proportions for using 2013 data:
- soy
- palm kernel
- oilcrops (groundnuts, sunflower seed, rape and mustardseed, sesame seed, cottonseed, other oilcrops = cakes)


We are using 2017 Food Balances data (prop going to feed) for:
- grains (barl, maiz, ocer, rice, sorg, whea, mil)
- pulses (peas, beans, other pulses)


General methods:
- download 2013 FAO Food Balances data
- download 2013 FAO Commodity Balances data


Soy methods:
- make sure all cake is going to feed (soy cake feed / soy cake production)
- calculate proportion soy DSQ going to processing (soy processing / soy DSQ)
- calculate proportion soy processing going to cake (feed) (soy cake production / (soybean processing))
- calculate proportion soy DSQ going to feed via processing (multiply previous 2 proportions)
- calculate proportion soy DSQ directly going to feed (soy feed / soy DSQ)
- add proportion soy directly going to feed to proportion soy DSQ going to feed via processing

NOTE:
Since we ended up doing soy cake production / soybean processing, these calculations could have been much simpler.
Essentiall all we're doing is (feed + cake)/ dsq of soybeans. But instead I used the more complicated route with all the multiplying proportions because that made sense when I was originally doing soy cake production / (soy cake + soy oil production) and multiplying that ratio by percent of dsq of soybeans that are processed. Since this is already done, I'm not going to change it, but something to keep in mind going forward and for explaining methods. It was also helpful doing it this way to check out proportions - like how much of soy processing is going to cake + oil, etc.


Load libraries
```{r}
library(tidyverse)
```

Data Paths
```{r}
raw_path <- "/home/shares/food-systems/social_justice_projects/food_trade_pressures/raw_data/"
FAO_path <- paste0(raw_path, "FAO/")
save_path <- "/home/shares/food-systems/social_justice_projects/food_trade_pressures/output_data/04a_calc_prop_feed_per_crop_country_FAO_FB_and_CB_2013/"
```


Read in FAO Food Balances data and subset to 2013
```{r}
fb <- read_csv(paste0(FAO_path, "FoodBalanceSheetsHistoric_E_All_Data_(Normalized).csv")) %>% 
  filter(Year == 2013)

# subset to only Unit == 1,000 tonnes
fb <- fb %>% 
  filter(Unit == "1000 tonnes")

# commodity balances are in tonnes so multiply values in fb by 1000 to get to real tonnes
fb <- fb %>% 
  mutate(value_tonnes = Value*1000)

# remove China
fb <- fb %>% 
  filter(!Area == "China")
```


Read in FAO Commodity Balances data and subset to 2013
```{r}
cb <- read_csv(paste0(FAO_path, "CommodityBalances_(non-food)_E_All_Data_(Normalized).csv")) %>% 
  filter(Year == 2013)
```


### Soy

Is all soy cake going to feed? Check in commodity balances data.
```{r}
# subset cb to soy cake
soy_cb <- cb %>% 
  filter(Item == "Soyabean Cake")

# subset and convert to wide
soy_cb_wide <- soy_cb %>% 
  select(Area, Element, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# calculate proportion DSQ of cake going to feed
soy_cb_cake_feed <- soy_cb_wide %>% 
  summarise(prop_cake_feed = Feed/`Domestic supply quantity`)
```


- calculate proportion soy DSQ going to processing (soy DSQ / soy processing)
- calculate proportion soy DSQ directly going to feed (soy feed / soy DSQ)
```{r}
# subset food balance data
soy_fb <- fb %>% 
  filter(Item == "Soyabeans") %>% 
  filter(Element ==  "Domestic supply quantity" | Element == "Feed" | Element == "Processing")

# pivot wider
soy_fb_wide <- soy_fb %>% 
  select(Area, Item, Element, value_tonnes) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes)

# calculate proportion soy going to processing and proportion soy going to feed
soy_fb_2 <- soy_fb_wide %>% 
  mutate(prop_process = Processing/`Domestic supply quantity`) %>% 
  mutate(prop_raw_feed = Feed/`Domestic supply quantity`)
```


- calculate proportion soy processing going to cake (feed) (soy cake production / soy processing)
```{r}
# rename production in cb to cake
soy_cb_2 <- soy_cb_wide %>% 
  rename(production_cake = Production) %>% 
  select(Area, production_cake)

# merge fb with cb data
soy_fb_cb <- full_join(soy_fb_2, soy_cb_2, by = "Area")

# subset oil 
soy_oil_fb <- fb %>% 
  filter(Item == "Soyabean Oil") %>% 
  filter(Element == "Production")
# only 3 countries have loss data for soybean oil production (USA, Japan and UAE)
# these losses are very small (3000 tonnes, 4000 tonnes and 141,000 tonnes for USA)

soy_oil_fb_2 <- soy_oil_fb %>% 
  select(Area, value_tonnes) %>% 
  rename(production_oil = value_tonnes)

# merge oil production
soy_fb_cb_2 <- full_join(soy_fb_cb, soy_oil_fb_2, by = "Area")

# calculate proportion of processed soy that goes to cake 
soy_fb_cb_3 <- soy_fb_cb_2 %>% 
  mutate(prop_cake_in_processing = production_cake/Processing)
# we want tonnes cake production / tonnes processing
# rather than divided by tonnes cake + oil production because there could be other things
# that are processed than just oil and soy, so we don't want to overestimate the amount
# going to feed. We're not accounting for losses in feed production, but we are assuming
# that those are small.

# convert Infinities to 1 (has cake production but pk processing = 0)
soy_fb_cb_3$prop_cake_in_processing[soy_fb_cb_3$prop_cake_in_processing > 100] <- 1  

# check to see if cake_oil_production is close to 100% of processing per country
# not quite 100% (see comments above)
soy_fb_cb_check <- soy_fb_cb_2 %>% 
  mutate(cake_oil_production = production_cake + production_oil) %>% 
  mutate(prop_cake_oil_in_processing = cake_oil_production/Processing)
```


- calculate proportion soy DSQ going to feed via processing (multiply previous 2 proportions)
```{r}
soy_fb_cb_4 <- soy_fb_cb_3 %>% 
  mutate(prop_feed = prop_process*prop_cake_in_processing)

# NaNs introduced where prop_process = NaN, which is where pk dsq = 0
# NAs introduced where processing = NA, so prop_process also = NA
```


- add proportion soy directly going to feed to proportion soy DSQ going to feed via processing
```{r}
# using rowwise to incorporate na.rm = TRUE, otherwise if either prop_feed or prop_raw_feed are NA, it 
# gives an NA for the prop_feed_total variable
soy_fb_cb_5 <- soy_fb_cb_4 %>% 
  rowwise() %>% 
  mutate(prop_feed_total = sum(prop_feed, prop_raw_feed, na.rm=TRUE))
# now all values that would be NA are 0 - so something to keep in mind with gapfilling

# save soy df
write_csv(soy_fb_cb_5, paste0(save_path, "perc_dsq_soy_used_as_feed.csv"))
```


For countries that are missing, we could use region averages, so we're keeping those in for now.



### Palm kernel
- palm oil comes from the palm oil fruit, palm kernel oil comes from the palm kernel (seed inside of fruit)
- only palm kernel cake is used as animal feed


Palm kernel methods:
- make sure all cake is going to feed
- calculate proportion palm kernel DSQ going to processing
- check to see if palm kernel oil production and palm kernel cake production are about equal to the weight of palm kernel dsq
- calculate proportion palm kernel processing going to cake (feed) 
- calculate proportion palm kernel DSQ going to feed via processing (multiply previous 2 proportions)
- check if any palm kernels are going to feed directly


Is all palm kernel cake going to feed?
- make sure all cake is going to feed
```{r}
# create df to with palm kernel cake 
pk_cake <- cb %>% 
  filter(Item == "Palmkernel Cake") %>% 
  select(Area, Element, Item, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# how much pk cake going to feed?
pk_cake_feed <- pk_cake %>% 
  summarise(perc_cake_to_feed = Feed/`Domestic supply quantity`)
# At least 96% of palm oil cake going to feed except in one country where it was 77%, which is likely an error
```


How much palm kernel is processed?
Not all of the DSQ is processed, and losses aren't always accounted for.
- calculate proportion palm kernel DSQ going to processing
```{r}
# subset palm kernels from commodity balances
pk_proc <- cb %>% 
  filter(Item == "Palm kernels") %>% 
  select(Area, Element, Item, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production"| Element == "Processing"| Element == "Losses") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# check if any palm kernels are going to feed directly
# no - only 3 tonnes of palm kernels going directly to feed and that's in Romania only, so ignore this

# calculate percent of palm kernel dsq going to processing - most of it is processed
# assume some loss when processing palm kernels into oil and cake (palm kernel includes kernels and palm fruit according to FAO definitions)
pk_process_perc <- pk_proc %>% 
  mutate(perc_pk_processed = Processing/`Domestic supply quantity`)

# I don't think palm kernels are going to anything besides palm kernel oil and cake
```


Does palm kernel oil and palm kernel cake production equal palm kernel DSQ?
Sum tonnes palm kernel cake production from CB with palm kernel oil production from FB and compare to palm kernel DSQ from CB (or from FB - same values, but FB is rounded to the thousands, so CB is more detailed)
- check to see if palm kernel oil production and palm kernel cake production are about equal to the weight of palm kernel dsq
```{r}
# create df with palm kernels
pk <- cb %>% 
  filter(Item == "Palm kernels") %>% 
  select(Area, Element, Item, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# create df with palm kernels oil
pk_oil <- fb %>% 
  filter(Item == "Palmkernel Oil") %>% 
  select(Area, Element, Item, value_tonnes) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes)

# combine pk, pk_cake, and pk_oil
pk_cake_prod <- pk_cake %>% 
  select(Area, Production) %>% 
  rename(pk_cake_production = Production)
pk_oil_prod <- pk_oil %>% 
  select(Area, Production) %>% 
  rename(pk_oil_production = Production)
pk_proc_2 <- pk_proc %>% 
  select(Area, `Domestic supply quantity`, Processing, Feed) %>% 
  rename(pk_dsq = `Domestic supply quantity`, pk_processing = Processing, pk_feed = Feed)
pk_1 <- full_join(pk_cake_prod, pk_oil_prod, by = "Area")
pk_2 <- full_join(pk_1, pk_proc_2, by = "Area")

# sum pk oil and pk cake
pk_2 <- pk_2 %>% 
  mutate(sum_cake_oil = (pk_cake_production + pk_oil_production))

# check to see if pk_cake_production + pk_oil_production is close to pk processing
# divide sum by dsq of pk
pk_2 <- pk_2 %>% 
  mutate(perc_proc_in_oil_cake = sum_cake_oil/pk_processing)
# palm kernel cake + oil account for about 80-100% of palm kernel processed tonnes

# only 50 countries produce palm kernel cake/ palm oil
# for most countries, over 85% of processed palm kernels are accounted for in the palm kernel cake and palm kernel oil production
# there are 11 countries with oil and cake accounting for <60% of processing tonnes
# we can assume those are lost or somehow else accounted for

# there is only 1 country (Romania) that include feed as a category for palm kernels, 
# and the amount is super low (3 tonnes)
```


- calculate proportion palm kernel processing going to cake (feed)
```{r}
# calculate proportion of processed palm kernels that goes to cake 
pk_2 <- pk_2 %>% 
  mutate(prop_cake_in_processing = pk_cake_production/pk_processing)

# subset previous df to only percent pk processing / dsq variable and Area
pk_process_perc_2 <- pk_process_perc %>% 
  select(Area, perc_pk_processed)

# join together
pk_3 <- full_join(pk_2, pk_process_perc_2, by = "Area")

# multiply prop of dsq that is processed by prop of processed that is turned into cake
pk_3 <- pk_3 %>% 
  mutate(prop_feed_total = prop_cake_in_processing*perc_pk_processed)

# add column that shows these are palm kernels
pk_3_clean <- pk_3 %>% 
  mutate(Item = "palm kernel")
# only 1 country includes feed (3 tonnes) for palm kernels

# save df
write_csv(pk_3_clean, paste0(save_path, "perc_dsq_palm_kernel_used_as_feed.csv"))
```


### Oilcrops

includes cake related to groundnuts, sunflower seed, rape and mustardseed, sesame seed, cottonseed, and other oilcrops

Oilcrop methods:
- make sure all cake is going to feed (cake feed / cake production)
- calculate proportion DSQ going to processing (processing / DSQ)
- calculate proportion processing going to cake (feed) (cake production / processing)
- calculate proportion DSQ going to feed via processing (multiply previous 2 proportions)
- calculate proportion DSQ directly going to feed (feed / DSQ)
- add proportion directly going to feed to proportion DSQ going to feed via processing



### Cottonseed

Is all cake going to feed? Check in commodity balances data. - Yes
```{r}
# subset cb to soy cake
cotton_cake_cb <- cb %>% 
  filter(Item == "Cottonseed Cake")

# subset and convert to wide
cotton_cake_cb_wide <- cotton_cake_cb %>% 
  select(Area, Element, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# calculate proportion DSQ of cake going to feed - pretty much 100% for all countries
cotton_cb_cake_feed <- cotton_cake_cb_wide %>% 
  mutate(prop_cake_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion cotton DSQ going to processing (cotton DSQ / cotton processing)
- calculate proportion cotton DSQ directly going to feed (cotton feed / cotton DSQ)
```{r}
# subset commodity balance data
cotton_cb <- cb %>% 
  filter(Item == "Cottonseed") %>% 
  filter(Element ==  "Domestic supply quantity" | Element == "Feed" | Element == "Processing")

# pivot wider
cotton_cb_wide <- cotton_cb %>% 
  select(Area, Item, Element, Value) %>% 
  pivot_wider(names_from = Element, values_from = Value)

# calculate proportion cotton going to processing and proportion cotton going to feed
cotton_cb_2 <- cotton_cb_wide %>% 
  mutate(prop_process = Processing/`Domestic supply quantity`) %>% 
  mutate(prop_raw_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion cotton processing going to cake (cotton cake production / cotton processing)
```{r}
# rename production in cb to cake
cotton_cake_cb_2 <- cotton_cake_cb_wide %>% 
  rename(production_cake = Production) %>% 
  select(Area, production_cake)

# merge fb with cb data
cotton_fb_cb <- full_join(cotton_cb_2, cotton_cake_cb_2, by = "Area")

# subset oil 
cotton_oil_fb <- fb %>% 
  filter(Item == "Cottonseed Oil") %>% 
  filter(Element == "Production")

cotton_oil_fb_2 <- cotton_oil_fb %>% 
  select(Area, value_tonnes) %>% 
  rename(production_oil = value_tonnes)

# merge oil production
cotton_fb_cb_2 <- full_join(cotton_fb_cb, cotton_oil_fb_2, by = "Area")

# calculate proportion of processed cotton that goes to cake 
cotton_fb_cb_3 <- cotton_fb_cb_2 %>% 
  mutate(prop_cake_in_processing = production_cake/Processing)

# convert Infinities to 1 (has cake production but processing = 0) - doesn't have infinities
# cotton_fb_cb_3$prop_cake_in_processing[cotton_fb_cb_3$prop_cake_in_processing > 100] <- 1  

# check to see if cake_oil_production is close to 100% of processing per country
# nope - seem to be lots of other products of the processing than cake + oil
cotton_fb_cb_check <- cotton_fb_cb_2 %>% 
  mutate(cake_oil_production = production_cake + production_oil) %>% 
  mutate(prop_cake_oil_in_processing = cake_oil_production/Processing)
```

- calculate proportion cotton DSQ going to feed via processing (multiply previous 2 proportions)
- add proportion cotton directly going to feed to proportion cotton DSQ going to feed via processing
```{r}
cotton_fb_cb_4 <- cotton_fb_cb_3 %>% 
  mutate(prop_feed = prop_process*prop_cake_in_processing)

cotton_fb_cb_5 <- cotton_fb_cb_4 %>% 
  rowwise() %>% 
  mutate(prop_feed_total = sum(prop_feed, prop_raw_feed, na.rm=TRUE))
# now all values that would be NA are 0 - so something to keep in mind with gapfilling

# save df
write_csv(cotton_fb_cb_5, paste0(save_path,"perc_dsq_cotton_used_as_feed.csv"))
```




### Groudnut

Is all cake going to feed? Check in commodity balances data. - Yes
```{r}
# subset cb to soy cake
nut_cb <- cb %>% 
  filter(Item == "Groundnut Cake")

# subset and convert to wide
nut_cb_wide <- nut_cb %>% 
  select(Area, Element, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# calculate proportion DSQ of cake going to feed - pretty much 100% for all countries
nut_cb_cake_feed <- nut_cb_wide %>% 
  mutate(prop_cake_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion nut DSQ going to processing (nut DSQ / nut processing)
- calculate proportion nut DSQ directly going to feed (nut feed / nut DSQ)
```{r}
# subset food balance data (groundnuts not in cb data, so have to use fb)
nut_fb <- fb %>% 
  filter(Item == "Groundnuts (Shelled Eq)") %>% 
  filter(Element ==  "Domestic supply quantity" | Element == "Feed" | Element == "Processing")

# pivot wider
nut_fb_wide <- nut_fb %>% 
  select(Area, Item, Element, value_tonnes) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes)

# calculate proportion nut going to processing and proportion nut going to feed
nut_fb_2 <- nut_fb_wide %>% 
  mutate(prop_process = Processing/`Domestic supply quantity`) %>% 
  mutate(prop_raw_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion nut processing going to cake (nut cake production / nut processing)
```{r}
# rename production in cb to cake
nut_cb_2 <- nut_cb_wide %>% 
  rename(production_cake = Production) %>% 
  select(Area, production_cake)

# merge fb with cb data
nut_fb_cb <- full_join(nut_fb_2, nut_cb_2, by = "Area")

# subset oil 
nut_oil_fb <- fb %>% 
  filter(Item == "Groundnut Oil") %>% 
  filter(Element == "Production")

nut_oil_fb_2 <- nut_oil_fb %>% 
  select(Area, value_tonnes) %>% 
  rename(production_oil = value_tonnes)

# merge oil production
nut_fb_cb_2 <- full_join(nut_fb_cb, nut_oil_fb_2, by = "Area")

# calculate proportion of processed nut that goes to cake 
nut_fb_cb_3 <- nut_fb_cb_2 %>% 
  mutate(prop_cake_in_processing = production_cake/Processing)

# convert Infinities to 1 (has cake production but processing = 0 b/c fb data is not as detailed as cb data so missing 
# values that are less than 1000 tonnes)
nut_fb_cb_3$prop_cake_in_processing[nut_fb_cb_3$prop_cake_in_processing > 100] <- 1  

# check to see if cake_oil_production is close to 100% of processing per country
# a high proportion of processing is accounted for by cake and oil, but seem to 
# be lots of other products of the processing than cake + oil
nut_fb_cb_check <- nut_fb_cb_2 %>% 
  mutate(cake_oil_production = production_cake + production_oil) %>% 
  mutate(prop_cake_oil_in_processing = cake_oil_production/Processing)
```

- calculate proportion nut DSQ going to feed via processing (multiply previous 2 proportions)
- add proportion soy directly going to feed to proportion soy DSQ going to feed via processing
```{r}
nut_fb_cb_4 <- nut_fb_cb_3 %>% 
  mutate(prop_feed = prop_process*prop_cake_in_processing)
# there is one instance where a country has a small amount of processing (1000 tonnes) but no dsq
# this country gets Inf for prop_feed

nut_fb_cb_5 <- nut_fb_cb_4 %>% 
  rowwise() %>% 
  mutate(prop_feed_total = sum(prop_feed, prop_raw_feed, na.rm=TRUE))
# one infinity value - convert to 1
nut_fb_cb_5$prop_feed_total[nut_fb_cb_5$prop_feed_total > 100] <- 1  

# now all values that would be NA are 0 - so something to keep in mind with gapfilling

# save df
write_csv(nut_fb_cb_5, paste0(save_path, "perc_dsq_groundnut_used_as_feed.csv"))
```





### Oilseed other
oilcrops, other in FB - lots of negative DSQ for this category in FAO

Is all cake going to feed? Check in commodity balances data. - Yes
```{r}
# subset cb to soy cake
other_oil_cb <- cb %>% 
  filter(Item == "Oilseed Cakes, Other")

# subset and convert to wide
other_oil_cb_wide <- other_oil_cb %>% 
  select(Area, Element, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# calculate proportion DSQ of cake going to feed - pretty much 100% for all countries (errors with negative dsq)
other_oil_cb_cake_feed <- other_oil_cb_wide %>% 
  mutate(prop_cake_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion other_oil DSQ going to processing (other_oil DSQ / other_oil processing)
- calculate proportion other_oil DSQ directly going to feed (other_oil feed / other_oil DSQ)
```{r}
# subset food balance data ("oilcrops other" not in cb data, so have to use fb)
other_oil_fb <- fb %>% 
  filter(Item == "Oilcrops, Other") %>% 
  filter(Element ==  "Domestic supply quantity" | Element == "Feed" | Element == "Processing")

# pivot wider
other_oil_fb_wide <- other_oil_fb %>% 
  select(Area, Item, Element, value_tonnes) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes)

# calculate proportion other_oil going to processing and proportion other_oil going to feed (decent amount going straight to feed)
other_oil_fb_2 <- other_oil_fb_wide %>% 
  mutate(prop_process = Processing/`Domestic supply quantity`) %>% 
  mutate(prop_raw_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion other_oil processing going to cake (other_oil cake production / other_oil processing)
```{r}
# rename production in cb to cake
other_oil_cb_2 <- other_oil_cb_wide %>% 
  rename(production_cake = Production) %>% 
  select(Area, production_cake)

# merge fb with cb data
other_oil_fb_cb <- full_join(other_oil_fb_2, other_oil_cb_2, by = "Area")

# subset oil 
other_oil_oil_fb <- fb %>% 
  filter(Item == "Groundnut Oil") %>% 
  filter(Element == "Production")

other_oil_oil_fb_2 <- other_oil_oil_fb %>% 
  select(Area, value_tonnes) %>% 
  rename(production_oil = value_tonnes)

# merge oil production
other_oil_fb_cb_2 <- full_join(other_oil_fb_cb, other_oil_oil_fb_2, by = "Area")

# calculate proportion of processed other_oil that goes to cake 
other_oil_fb_cb_3 <- other_oil_fb_cb_2 %>% 
  mutate(prop_cake_in_processing = production_cake/Processing)

# convert Infinities to 1 (has cake production but processing = 0 b/c fb data is not as detailed as cb data so missing 
# values that are less than 1000 tonnes)
other_oil_fb_cb_3$prop_cake_in_processing[other_oil_fb_cb_3$prop_cake_in_processing > 100] <- 1  

# check to see if cake_oil_production is close to 100% of processing per country
# a high proportion of processing is accounted for by cake and oil, but seem to 
# be lots of other products of the processing than cake + oil
other_oil_fb_cb_check <- other_oil_fb_cb_2 %>% 
  mutate(cake_oil_production = production_cake + production_oil) %>% 
  mutate(prop_cake_oil_in_processing = cake_oil_production/Processing)
```

- calculate proportion other_oil DSQ going to feed via processing (multiply previous 2 proportions)
- add proportion soy directly going to feed to proportion soy DSQ going to feed via processing
```{r}
other_oil_fb_cb_4 <- other_oil_fb_cb_3 %>% 
  mutate(prop_feed = prop_process*prop_cake_in_processing)
# there is one instance where a country has a small amount of processing (1000 tonnes) but no dsq
# this country gets Inf for prop_feed

other_oil_fb_cb_5 <- other_oil_fb_cb_4 %>% 
  rowwise() %>% 
  mutate(prop_feed_total = sum(prop_feed, prop_raw_feed, na.rm=TRUE))

# one infinity value - convert to 1
other_oil_fb_cb_5$prop_feed_total[other_oil_fb_cb_5$prop_feed_total > 100] <- 1  
# now all values that would be NA are 0 - so something to keep in mind with gapfilling

# save df
write_csv(other_oil_fb_cb_5, paste0(save_path, "perc_dsq_other_oil_used_as_feed.csv"))
```






### Rape and mustardseed

Is all cake going to feed? Check in commodity balances data. - Yes
```{r}
# subset cb to soy cake
mustard_cake_cb <- cb %>% 
  filter(Item == "Rape and Mustard Cake")

# subset and convert to wide
mustard_cake_cb_wide <- mustard_cake_cb %>% 
  select(Area, Element, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# calculate proportion DSQ of cake going to feed - pretty much 100% for all countries
mustard_cb_cake_feed <- mustard_cake_cb_wide %>% 
  mutate(prop_cake_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion mustard DSQ going to processing (mustard DSQ / mustard processing)
- calculate proportion mustard DSQ directly going to feed (mustard feed / mustard DSQ)
```{r}
# subset commodity balance data
mustard_cb <- cb %>% 
  filter(Item == "Rape and Mustardseed") %>% 
  filter(Element ==  "Domestic supply quantity" | Element == "Feed" | Element == "Processing")

# pivot wider
mustard_cb_wide <- mustard_cb %>% 
  select(Area, Item, Element, Value) %>% 
  pivot_wider(names_from = Element, values_from = Value)

# calculate proportion mustard going to processing and proportion mustard going to feed
mustard_cb_2 <- mustard_cb_wide %>% 
  mutate(prop_process = Processing/`Domestic supply quantity`) %>% 
  mutate(prop_raw_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion mustard processing going to cake (mustard cake production / mustard processing)
```{r}
# rename production in cb to cake
mustard_cake_cb_2 <- mustard_cake_cb_wide %>% 
  rename(production_cake = Production) %>% 
  select(Area, production_cake)

# merge fb with cb data
mustard_fb_cb <- full_join(mustard_cb_2, mustard_cake_cb_2, by = "Area")

# subset oil 
mustard_oil_fb <- fb %>% 
  filter(Item == "Rape and Mustard Oil") %>% 
  filter(Element == "Production")

mustard_oil_fb_2 <- mustard_oil_fb %>% 
  select(Area, value_tonnes) %>% 
  rename(production_oil = value_tonnes)

# merge oil production
mustard_fb_cb_2 <- full_join(mustard_fb_cb, mustard_oil_fb_2, by = "Area")

# calculate proportion of processed mustard that goes to cake 
mustard_fb_cb_3 <- mustard_fb_cb_2 %>% 
  mutate(prop_cake_in_processing = production_cake/Processing)

# check to see if cake_oil_production is close to 100% of processing per country
# between 80 - 100% for most countries
mustard_fb_cb_check <- mustard_fb_cb_2 %>% 
  mutate(cake_oil_production = production_cake + production_oil) %>% 
  mutate(prop_cake_oil_in_processing = cake_oil_production/Processing)
```

- calculate proportion mustard DSQ going to feed via processing (multiply previous 2 proportions)
- add proportion mustard directly going to feed to proportion mustard DSQ going to feed via processing
```{r}
mustard_fb_cb_4 <- mustard_fb_cb_3 %>% 
  mutate(prop_feed = prop_process*prop_cake_in_processing)

mustard_fb_cb_5 <- mustard_fb_cb_4 %>% 
  rowwise() %>% 
  mutate(prop_feed_total = sum(prop_feed, prop_raw_feed, na.rm=TRUE))
# now all values that would be NA are 0 - so something to keep in mind with gapfilling

# save df
write_csv(mustard_fb_cb_5, paste0(save_path, "perc_dsq_rapemustard_used_as_feed.csv"))
```





### Sesameseed

Is all cake going to feed? Check in commodity balances data. - Yes
```{r}
# subset cb to soy cake
sesame_cb <- cb %>% 
  filter(Item == "Sesameseed Cake")

# subset and convert to wide
sesame_cb_wide <- sesame_cb %>% 
  select(Area, Element, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# calculate proportion DSQ of cake going to feed - all 100%
sesame_cb_cake_feed <- sesame_cb_wide %>% 
  mutate(prop_cake_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion sesame DSQ going to processing (sesame DSQ / sesame processing)
- calculate proportion sesame DSQ directly going to feed (sesame feed / sesame DSQ)
```{r}
# subset food balance data (sesameseeds not in cb data, so have to use fb - less detailed so will introduce error)
sesame_fb <- fb %>% 
  filter(Item == "Sesame seed") %>% 
  filter(Element ==  "Domestic supply quantity" | Element == "Feed" | Element == "Processing")

# pivot wider
sesame_fb_wide <- sesame_fb %>% 
  select(Area, Item, Element, value_tonnes) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes)

# calculate proportion sesame going to processing and proportion sesame going to feed
sesame_fb_2 <- sesame_fb_wide %>% 
  mutate(prop_process = Processing/`Domestic supply quantity`) %>% 
  mutate(prop_raw_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion sesame processing going to cake (sesame cake production / sesame processing)
```{r}
# rename production in cb to cake
sesame_cb_2 <- sesame_cb_wide %>% 
  rename(production_cake = Production) %>% 
  select(Area, production_cake)

# merge fb with cb data
sesame_fb_cb <- full_join(sesame_fb_2, sesame_cb_2, by = "Area")

# subset oil 
sesame_oil_fb <- fb %>% 
  filter(Item == "Sesameseed Oil") %>% 
  filter(Element == "Production")

sesame_oil_fb_2 <- sesame_oil_fb %>% 
  select(Area, value_tonnes) %>% 
  rename(production_oil = value_tonnes)

# merge oil production
sesame_fb_cb_2 <- full_join(sesame_fb_cb, sesame_oil_fb_2, by = "Area")

# calculate proportion of processed sesame that goes to cake 
sesame_fb_cb_3 <- sesame_fb_cb_2 %>% 
  mutate(prop_cake_in_processing = production_cake/Processing)

# convert Infinities to 1 (has cake production but processing = 0 b/c fb data is not as detailed as cb data so missing 
# values that are less than 1000 tonnes)
sesame_fb_cb_3$prop_cake_in_processing[sesame_fb_cb_3$prop_cake_in_processing > 100] <- 1  

# check to see if cake_oil_production is close to 100% of processing per country
# a high proportion of processing (80-100%) is accounted for by cake and oil
sesame_fb_cb_check <- sesame_fb_cb_2 %>% 
  mutate(cake_oil_production = production_cake + production_oil) %>% 
  mutate(prop_cake_oil_in_processing = cake_oil_production/Processing)
```

- calculate proportion sesame DSQ going to feed via processing (multiply previous 2 proportions)
- add proportion soy directly going to feed to proportion soy DSQ going to feed via processing
```{r}
sesame_fb_cb_4 <- sesame_fb_cb_3 %>% 
  mutate(prop_feed = prop_process*prop_cake_in_processing)

sesame_fb_cb_5 <- sesame_fb_cb_4 %>% 
  rowwise() %>% 
  mutate(prop_feed_total = sum(prop_feed, prop_raw_feed, na.rm=TRUE))
# now all values that would be NA are 0 - so something to keep in mind with gapfilling

# save df
write_csv(sesame_fb_cb_5, paste0(save_path, "perc_dsq_sesame_used_as_feed.csv"))
```







### Sunflower seed

Is all cake going to feed? Check in commodity balances data. - Yes
```{r}
# subset cb to soy cake
sunflower_cb <- cb %>% 
  filter(Item == "Sunflowerseed Cake")

# subset and convert to wide
sunflower_cb_wide <- sunflower_cb %>% 
  select(Area, Element, Value) %>% 
  filter(Element == "Feed" | Element == "Domestic supply quantity" | Element == "Production") %>% 
  pivot_wider(names_from = Element, values_from = Value)

# calculate proportion DSQ of cake going to feed - mostly all 99 - 100%
sunflower_cb_cake_feed <- sunflower_cb_wide %>% 
  mutate(prop_cake_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion sunflower DSQ going to processing (sunflower DSQ / sunflower processing)
- calculate proportion sunflower DSQ directly going to feed (sunflower feed / sunflower DSQ)
```{r}
# subset food balance data (sunflowerseeds not in cb data, so have to use fb - less detailed so will introduce error)
sunflower_fb <- fb %>% 
  filter(Item == "Sunflower seed") %>% 
  filter(Element ==  "Domestic supply quantity" | Element == "Feed" | Element == "Processing")

# pivot wider
sunflower_fb_wide <- sunflower_fb %>% 
  select(Area, Item, Element, value_tonnes) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes)

# calculate proportion sunflower going to processing and proportion sunflower going to feed
sunflower_fb_2 <- sunflower_fb_wide %>% 
  mutate(prop_process = Processing/`Domestic supply quantity`) %>% 
  mutate(prop_raw_feed = Feed/`Domestic supply quantity`)
```

- calculate proportion sunflower processing going to cake (sunflower cake production / sunflower processing)
```{r}
# rename production in cb to cake
sunflower_cb_2 <- sunflower_cb_wide %>% 
  rename(production_cake = Production) %>% 
  select(Area, production_cake)

# merge fb with cb data
sunflower_fb_cb <- full_join(sunflower_fb_2, sunflower_cb_2, by = "Area")

# subset oil 
sunflower_oil_fb <- fb %>% 
  filter(Item == "Sunflowerseed Oil") %>% 
  filter(Element == "Production")

sunflower_oil_fb_2 <- sunflower_oil_fb %>% 
  select(Area, value_tonnes) %>% 
  rename(production_oil = value_tonnes)

# merge oil production
sunflower_fb_cb_2 <- full_join(sunflower_fb_cb, sunflower_oil_fb_2, by = "Area")

# calculate proportion of processed sunflower that goes to cake 
sunflower_fb_cb_3 <- sunflower_fb_cb_2 %>% 
  mutate(prop_cake_in_processing = production_cake/Processing)

# convert Infinities to 1 (has cake production but processing = 0 b/c fb data is not as detailed as cb data so missing 
# values that are less than 1000 tonnes)
sunflower_fb_cb_3$prop_cake_in_processing[sunflower_fb_cb_3$prop_cake_in_processing > 100] <- 1  

# check to see if cake_oil_production is close to 100% of processing per country
# a high proportion of processing (70-100%) is accounted for by cake and oil
sunflower_fb_cb_check <- sunflower_fb_cb_2 %>% 
  mutate(cake_oil_production = production_cake + production_oil) %>% 
  mutate(prop_cake_oil_in_processing = cake_oil_production/Processing)
```

- calculate proportion sunflower DSQ going to feed via processing (multiply previous 2 proportions)
- add proportion soy directly going to feed to proportion soy DSQ going to feed via processing
```{r}
sunflower_fb_cb_4 <- sunflower_fb_cb_3 %>% 
  mutate(prop_feed = prop_process*prop_cake_in_processing)

sunflower_fb_cb_5 <- sunflower_fb_cb_4 %>% 
  rowwise() %>% 
  mutate(prop_feed_total = sum(prop_feed, prop_raw_feed, na.rm=TRUE))
# now all values that would be NA are 0 - so something to keep in mind with gapfilling

# save df
write_csv(sunflower_fb_cb_5, paste0(save_path, "perc_dsq_sunflower_used_as_feed.csv"))
```



Combine all oilcrops (not soy or palm kernel) and calculate a total proportion feed for all oil crops
sum feed + sum cake / sum dsq
```{r}
# read in oilcrops percent feed csvs
sunflower <- read_csv(paste0(save_path, "perc_dsq_sunflower_used_as_feed.csv"))
sesame <- read_csv(paste0(save_path, "perc_dsq_sesame_used_as_feed.csv"))
rapemustard <- read_csv(paste0(save_path, "perc_dsq_rapemustard_used_as_feed.csv"))
other_oil <- read_csv(paste0(save_path, "perc_dsq_other_oil_used_as_feed.csv"))
groundnut <- read_csv(paste0(save_path, "perc_dsq_groundnut_used_as_feed.csv"))
cotton <- read_csv(paste0(save_path, "perc_dsq_cotton_used_as_feed.csv"))

# rename columns and subset to columns of interest
rename_subset_oilcrops <- function(df, dsq_col_name, feed_col_name, cake_col_name, prop_feed_col_name){
df <- df %>% 
  select(Area, `Domestic supply quantity`, Feed, production_cake, prop_feed_total) %>% 
  rename({{dsq_col_name}} := `Domestic supply quantity`, {{feed_col_name}} := Feed, {{cake_col_name}} := production_cake, {{prop_feed_col_name}} := prop_feed_total)  
}

sunflower_1 <- rename_subset_oilcrops(sunflower, sunflower_dsq, sunflower_feed, sunflower_cake, sunflower_prop_feed)
cotton_1 <- rename_subset_oilcrops(cotton, cotton_dsq, cotton_feed, cotton_cake, cotton_prop_feed)
groundnut_1 <- rename_subset_oilcrops(groundnut, groundnut_dsq, groundnut_feed, groundnut_cake, groundnut_prop_feed)
other_oil_1 <- rename_subset_oilcrops(other_oil, other_oil_dsq, other_oil_feed, other_oil_cake, other_oil_prop_feed)
rapemustard_1 <- rename_subset_oilcrops(rapemustard, rapemustard_dsq, rapemustard_feed, rapemustard_cake, rapemustard_prop_feed)
sesame_1 <- rename_subset_oilcrops(sesame, sesame_dsq, sesame_feed, sesame_cake, sesame_prop_feed)

# combine dfs
oils <- sunflower_1 %>% 
  full_join(cotton_1, by = "Area") %>% 
  full_join(groundnut_1, by = "Area") %>% 
  full_join(other_oil_1, by = "Area") %>% 
  full_join(rapemustard_1, by = "Area") %>% 
  full_join(sesame_1, by = "Area")

# sum dsq, feed, and cake, calculate proportion going to feed
oils <- oils %>% 
  rowwise() %>% 
  mutate(sum_dsq = sum(sunflower_dsq, cotton_dsq, groundnut_dsq, other_oil_dsq, rapemustard_dsq, sesame_dsq, na.rm = TRUE)) %>% 
  mutate(sum_feed = sum(sunflower_feed, cotton_feed, groundnut_feed, other_oil_feed, rapemustard_feed, sesame_feed, na.rm = TRUE)) %>% 
  mutate(sum_cake = sum(sunflower_cake, cotton_cake, groundnut_cake, other_oil_cake, rapemustard_cake, sesame_cake, na.rm = TRUE)) %>% 
  mutate(oil_prop_feed = (sum_feed + sum_cake)/sum_dsq)

# one country where oil_prop_feed = Inf. Has cake production, but no dsq. 
# Also 8 cases where countries oil_prop_feed > 1 because more cake production than dsq.
# convert all of these cases to 1.
oils <- oils %>% 
  mutate(oil_prop_feed = case_when(oil_prop_feed > 1 ~ 1,
                                   TRUE ~ oil_prop_feed))

# there are 2 cases of NaNs for oil_prop_feed where there dsq, feed, and cake all = 0.
# we'll keep these as NaNs to be gapfilled later.

# save csv
write_csv(oils, paste0(save_path, "perc_dsq_oilcrops_used_as_feed.csv"))
```
