==> 04b_gapfill_perc_feed_soy_palm_oilcrops_2013_CB_FB.Rmd <==
---
title: "Gapfill percent dsq used as feed with regional data - oilcrops, soy, palm kernels"
author: "Haley Epperly"
date: "2022-11-04"
output: html_document
editor_options: 
  chunk_output_type: console
---

Objective: Gapfill missing (NA) and 0s in the percent domestic supply quantity used as feed for aggregate oil crops (sunflowerseed, sesameseed, rape and mustardseed, cottonseed, groundnuts), soy, and palm kernels. For each country, missing data will be filled with regional feed estimates from the 2013 FAO Food Balances and Commodity Balances data. We already calculated regional feed estimates in script calc_prop_feed_per_crop_country_FAO_FB_and_CB_2013. 
If regional values are NA, then fill with the world value. If regional values are 0, then keep the country level 0.

These data that need to be gapfilled are located here following this naming convention:
perc_dsq_soy_used_as_feed
food_trade_ej/scripts/feed_in_trade_matrix/outputs

This is where the gapfilled data will be stored following this naming convention:
perc_dsq_soy_used_as_feed_gapfilled
food_trade_ej/scripts/feed_in_trade_matrix/outputs

After gapfilling these crops, we will combine all gapfilled feed dfs (including grains and pulses) into one df in another script.


Load libraries
```{r}
library(tidyverse)
library(countrycode)
```

Data Paths
```{r}
raw_path <- "/home/shares/food-systems/social_justice_projects/food_trade_pressures/raw_data/"
FAO_path <- paste0(raw_path, "FAO/")
kastner_path <- paste0(raw_path, "kastner_2022/")
fish_trade_path <- paste0(raw_path, "trade/aquatic_food_trade/")
feed_4a_path <- "/home/shares/food-systems/social_justice_projects/food_trade_pressures/output_data/04a_calc_prop_feed_per_crop_country_FAO_FB_and_CB_2013/"
save_path <- "/home/shares/food-systems/social_justice_projects/food_trade_pressures/output_data/04b_gapfill_perc_feed_soy_palm_oilcrops_2013_CB_FB/"
```


Need to add in missing countries that aren't even rows in these data
Combination of FAO country names and additional countries from fisheries trade data.
We won't need all of these countries, but it's good to include them all just in case.
```{r}
# rather than reading in the whole FAO country list, we'll just consider countries included in trade matrices output
# read in FAO country list
tm <- read_csv(paste0(kastner_path, "trade_matrix_by_country_of_origin_2017_import.csv"))

tm_1 <- tm %>% 
  select(Consumer_Country_Code) %>% 
  unique() %>% 
  mutate(country = countrycode(Consumer_Country_Code, "fao", "country.name"))

tm_2 <- tm_1 %>% 
  mutate(country = case_when(Consumer_Country_Code == 41 ~ "China, mainland",
                             Consumer_Country_Code == 299 ~ "Palestine",
                             TRUE ~ country))

# read in fisheries trade data 
fish_tm <- read_csv(paste0(fish_trade_path, "20230420_nceas_consumption.csv"))

# add country names and subset
fish_tm_1 <- fish_tm %>% 
  select(consumer_iso3c) %>% 
  unique() %>% 
  mutate(country = countrycode(consumer_iso3c, 'iso3c', 'country.name')) %>% 
  filter(!is.na(country)) # remove one NA (iso3c = NEI, Netherlands East Indies, not in FAO data either)


# combine dfs to get complete list of countries
countries <- full_join(tm_2, fish_tm_1, by = "country")

# add iso3c
countries_1 <- countries %>% 
  mutate(iso3 = countrycode(country, "country.name", "iso3c")) %>% 
  select(country, iso3)
```



Read in previously created csvs of percent feed for soy, add region column, and gapfill
```{r}
soy <- read_csv(paste0(feed_4a_path, "perc_dsq_soy_used_as_feed.csv"))

# have to rename Turkey because it got a ? symbol added when saved as csv previously
# can't use %in% or == to identify it because it isn't showing up as a string
soy_1 <- soy
soy_1[152,1]                    # check to make sure this is Turkey
soy_1[152,1] <- "Turkey"
soy_1[37,1]                     # check to make sure this is Côte d'Ivoire
soy_1[37,1] <- "Côte d'Ivoire"
soy_1[33,1]                     # check to make sure this is Taiwan
soy_1[33,1] <- "Taiwan"

# remove item column and rename prop_feed 
# also have to do this line with the utf coding to avoid error when adding regions with countrycode package
soy_2 <- soy_1 %>% 
  select(Area, `Domestic supply quantity`, Feed, production_cake, prop_feed_total) %>% 
  rename(soy_dsq = `Domestic supply quantity`, 
         soy_feed = Feed, 
         soy_cake = production_cake, 
         soy_prop_feed = prop_feed_total) %>% 
  mutate_if(is.character, utf8::utf8_encode)

# add iso3 column
soy_a <- soy_2 %>% 
  mutate(iso3 = countrycode(Area, "country.name", "iso3c"))

# join with country list
soy_b <- soy_a %>% 
  full_join(countries_1, by = "iso3")

# fill in country name where NA
soy_c <- soy_b %>% 
  mutate(Area = case_when(is.na(Area) ~ country, 
                          TRUE ~ Area)) %>% 
  select(-country)

# check iso3 codes for duplicates
# remove duplicates: China, China, mainland, Palestinian Territories
# this gets rid of duplicates of mainland China
soy_d <- soy_c %>% 
  unique()

# remove Palestinian Territories because we have Palestine
soy_2 <- soy_d %>% 
  filter(!Area == "Palestinian Territories")

# add region
# we need a combination of UN subregion and intermediate codes
# add subregion

soy_3 <- soy_2 %>% 
  mutate(subregion = countrycode(iso3, 'iso3c', 'un.regionsub.name')) %>% 
  mutate(intermed_region = countrycode(iso3, 'iso3c', 'un.regionintermediate.name')) %>% 
  mutate(region = case_when(is.na(intermed_region) ~ subregion,
                            TRUE ~ intermed_region))  
# more countries are matched with region when we use country name instead of iso3 codes

# fill in Taiwan 
soy_3$region[soy_3$Area == "Taiwan"] <- "Eastern Asia"

# for the previously calculated regions (bottom rows in df) that show up as NA 
# when trying to add the region column with the countrycode package,
# copy their region name into the region column
soy_4 <- soy_3 %>% 
  mutate(region = case_when(is.na(region) ~ Area, TRUE ~ region))

# remove extra columns and reorder
soy_5 <- soy_4 %>% 
  select(-c(subregion, intermed_region)) %>% 
  select(Area, iso3, region, soy_dsq, soy_feed, soy_cake, soy_prop_feed) # reorder

# replace NAs and 0s with region values (0s include NAs because of how this was previously calculated)
soy_6 <- soy_5 %>% 
  group_by(region) %>% 
  mutate(soy_prop_feed_gf = case_when(soy_prop_feed == 0 ~ soy_prop_feed[Area == region],
         TRUE ~ soy_prop_feed)) %>% 
  mutate(soy_prop_feed_gf = case_when(is.na(soy_prop_feed_gf) ~ soy_prop_feed[Area == region],
         TRUE ~ soy_prop_feed_gf)) %>% 
  ungroup()

# save csv
write_csv(soy_6, paste0(save_path, "perc_dsq_soy_used_as_feed_gapfilled.csv"))
```



Read in previously created csvs of percent feed for palm kernels, add region column, and gapfill
```{r}
pk <- read_csv(paste0(feed_4a_path, "perc_dsq_palm_kernel_used_as_feed.csv"))

# have to rename Turkey because it got a ? symbol added when saved as csv previously
# can't use %in% or == to identify it because it isn't showing up as a string
pk_1 <- pk
pk_1[105,1]                    # check that this is Turkey
pk_1[105,1] <- "Turkey"
pk_1[27,1]                     # check that this is Côte d'Ivoire
pk_1[27,1] <- "Côte d'Ivoire"
pk_1[23,1]                     # check that this is Taiwan
pk_1[23,1] <- "Taiwan"

# add row for Polynesia
pk_1 <- pk_1 %>% 
  add_row(Area = "Polynesia")

# remove item column and rename prop_feed 
# also have to do this line with the utf coding to avoid error when adding regions with countrycode package
pk_2 <- pk_1 %>% 
  select(Area, pk_dsq, pk_feed, pk_cake_production, prop_feed_total) %>% 
  rename(pk_cake = pk_cake_production, pk_prop_feed = prop_feed_total) %>% 
  mutate_if(is.character, utf8::utf8_encode)

# add iso3 column
pk_a <- pk_2 %>% 
  mutate(iso3 = countrycode(Area, "country.name", "iso3c"))

# join with country list
pk_b <- pk_a %>% 
  full_join(countries_1, by = "iso3")

# fill in country name where NA
pk_c <- pk_b %>% 
  mutate(Area = case_when(is.na(Area) ~ country, 
                          TRUE ~ Area)) %>% 
  select(-country)

# check iso3 codes for duplicates
# remove duplicates: China, China, mainland, Palestinian Territories
# this gets rid of duplicates of mainland China
pk_d <- pk_c %>% 
  unique()

# remove Palestinian Territories because we have Palestine
pk_2 <- pk_d %>% 
  filter(!Area == "Palestinian Territories")

# add region
# we need a combination of UN subregion and intermediate codes
# add subregion

pk_3 <- pk_2 %>% 
  mutate(subregion = countrycode(iso3, 'iso3c', 'un.regionsub.name')) %>% 
  mutate(intermed_region = countrycode(iso3, 'iso3c', 'un.regionintermediate.name')) %>% 
  mutate(region = case_when(is.na(intermed_region) ~ subregion,
                            TRUE ~ intermed_region))  
# more countries are matched with region when we use country name instead of iso3 codes

# fill in China Taiwan
pk_3$region[pk_3$Area == "Taiwan"] <- "Eastern Asia"

# for the previously calculated regions that show up as NA 
# when trying to add the region column with the countrycode package,
# copy their region name into the region column
pk_4 <- pk_3 %>% 
  mutate(region = case_when(is.na(region) ~ Area, TRUE ~ region))

# remove extra columns and reorder
pk_5 <- pk_4 %>% 
  select(-c(subregion, intermed_region)) %>% 
  select(Area, iso3, region, pk_dsq, pk_feed, pk_cake, pk_prop_feed) # reorder

# replace 0s and NAs with region values (0s include NAs because of how this was previously calculated)
pk_6 <- pk_5 %>% 
  group_by(region) %>% 
  mutate(pk_prop_feed_gf = case_when(pk_prop_feed == 0 ~ pk_prop_feed[Area == region],
                            is.na(pk_prop_feed) ~ pk_prop_feed[Area == region],
                            TRUE ~ pk_prop_feed)) %>%
  ungroup()

# we don't have region values for the following regions (they also have super low dsq):
# Australia and NZ
# Central Asia
# Melanesia
# Northern Africa
# Northern America
# Northern Europe
# Oceania
# Southern Africa
# Southern Asia
# Western Asia

# fill in missing regional values with continent level values

# add continent to df
pk_6$continent <- countrycode(pk_6$Area, origin = 'country.name', destination = 'continent')

# fill in continents names for gapfilling
# countrycode function doesn't work because you are not allowed to set continent
# or un.regionsub.name as the origin variable
pk_7 <- pk_6 %>% 
  mutate(continent = case_when(Area == "Africa" ~ "Africa",
                               Area == "Europe" ~ "Europe",
                               Area == "Americas" ~ "Americas",
                               Area == "Oceania" ~ "Oceania",
                               Area == "Asia" ~ "Asia",
                               Area == "Africa" ~ "Africa",
                               TRUE ~ continent))

# Oceania doesn't have a value for percent feed, so replace with World value
pk_8 <- pk_7 %>% 
  mutate(pk_prop_feed_gf = case_when(Area == "Oceania" ~ 0.5225788,
                                     TRUE ~ pk_prop_feed_gf))


# replace NAs with continent values 
pk_9 <- pk_8 %>% 
  group_by(continent) %>% 
  mutate(pk_prop_feed_gf_2 = case_when(is.na(pk_prop_feed_gf) ~ pk_prop_feed_gf[Area == continent],
         TRUE ~ pk_prop_feed_gf)) %>% 
  ungroup()

# some regions have NAs (e.g., Australia and New Zealand) that's fine though because all countries
# have been gapfilled. We'll keep the NAs for regions in. No percent feed values below 0 or above 1.

# clean csv
pk_10 <- pk_9 %>% 
  select(-c(pk_prop_feed_gf, continent))

# save csv
write_csv(pk_10, paste0(save_path, "perc_dsq_palm_kernel_used_as_feed_gapfilled.csv"))
```


Read in previously created csvs of percent feed for oil crops, add region column, and gapfill the aggreagte oil crop column
Not going to gapfill each individual oil crop for now, but can come back and do that if we need it later.
Only 3 NAs at country level for percent aggregate oil crops going to feed.
```{r}
oil <- read_csv(paste0(feed_4a_path, "perc_dsq_oilcrops_used_as_feed.csv"))

# have to rename Turkey because it got a ? symbol added when saved as csv previously
# can't use %in% or == to identify it because it isn't showing up as a string
oil_1 <- oil
oil_1[131,1]                    # make sure this is Turkey
oil_1[131,1] <- "Turkey"
oil_1[184,1]                    # make sure this is Côte d'Ivoire
oil_1[184,1] <- "Côte d'Ivoire"
oil_1[33,1]                     # make sure this is Taiwan
oil_1[33,1] <- "Taiwan"

# have to do this line with the utf coding to avoid error when adding regions with countrycode package
oil_2 <- oil_1 %>% 
  mutate_if(is.character, utf8::utf8_encode)

# add iso3 column
oil_a <- oil_2 %>% 
  mutate(iso3 = countrycode(Area, "country.name", "iso3c"))

# join with country list
oil_b <- oil_a %>% 
  full_join(countries_1, by = "iso3")

# fill in country name where NA
oil_c <- oil_b %>% 
  mutate(Area = case_when(is.na(Area) ~ country, TRUE ~ Area)) %>% 
  select(-country)

# check iso3 codes for duplicates
# remove duplicates: China, China, mainland, Palestinian Territories
# this gets rid of duplicates of mainland China
oil_d <- oil_c %>% 
  unique()

# remove Palestinian Territories because we have Palestine
oil_2 <- oil_d %>% 
  filter(!Area == "Palestinian Territories")

# add region
# we need a combination of UN subregion and intermediate codes
# add subregion

oil_3 <- oil_2 %>% 
  mutate(subregion = countrycode(iso3, 'iso3c', 'un.regionsub.name')) %>% 
  mutate(intermed_region = countrycode(iso3, 'iso3c', 'un.regionintermediate.name')) %>% 
  mutate(region = case_when(is.na(intermed_region) ~ subregion,
                            TRUE ~ intermed_region))  
# more countries are matched with region when we use country name instead of iso3 codes

# fill in Taiwan
oil_3$region[oil_3$Area == "Taiwan"] <- "Eastern Asia"

# for the previously calculated regions that show up as NA 
# when trying to add the region column with the countrycode package,
# copy their region name into the region column
oil_4 <- oil_3 %>% 
  mutate(region = case_when(is.na(region) ~ Area, TRUE ~ region))

# remove extra columns and reorder
oil_5 <- oil_4 %>% 
  select(-c(subregion, intermed_region))
oil_5 <- oil_5[,c(1, 30:31, 2:29)]

# replace 0s and NAs with region values
oil_6 <- oil_5 %>% 
  group_by(region) %>% 
  mutate(oil_prop_feed_gf = case_when(oil_prop_feed == 0 ~ oil_prop_feed[Area == region],
                                      is.na(oil_prop_feed) ~ oil_prop_feed[Area == region],
         TRUE ~ oil_prop_feed))  %>%
  ungroup()

# Melanesia, Micronesia, and Polynesia all have 0 for % feed of oilcrops going to feed
# these countries only grow some rape and mustardseed and some groundnuts
# we'll keep this as 0 - maybe their "cake" goes to something else like flour

# save csv
write_csv(oil_6, paste0(save_path, "perc_dsq_oilcrops_used_as_feed_gapfilled.csv"))
```
