==> 04c_calc_prop_feed_pulses_grains_2017_FAO_FB.Rmd <==
---
title: "Calculating proportion crop used as feed using 2017 FAO food balance data"
author: "Haley Epperly"
date: "2022-11-01"
output: html_document
editor_options: 
  chunk_output_type: console
---

Objective: Calculate the proportion of each country's domestic supply quantity that they used as feed for each relevant crop using 2017 FAO Food Balance data.

Check: USDA says the USA uses ~40% of their corn for feed, which is in line with calculations using 2017 FAO FB data and dividing feed by domestic supply quantity for maize (41.6%).
Source: https://www.ers.usda.gov/topics/crops/corn-and-other-feedgrains/feedgrains-sector-at-a-glance/#:~:text=Corn%20is%20a%20major%20component,of%20total%20domestic%20corn%20use.

Steps:
1. Download FAO FB data subset to 2017
2. For each crop and country, divide the tonnes in feed by tonnes domestic supply quantity
3. Combine dfs for all grains
4. Download FAO FB data with regions
5. Merge grains df with region by country name (countrycode package)
6. Gapfill missing data with region data
7. Save grains output
8. Repeat for pulses

Grains (and name in FAO FB data [2010-]):
- barl - barley (FAO = Barley and products)
- maiz -  maize (FAO = Maize and products)
- sorg- sorghum (FAO = Sorghum and products)
- xmil - millet (FAO = Millet and products)
- rice - rice (FAO = Rice and products)
- whea - wheat (FAO = Wheat and products)
- ocer (other cereals) includes FAO categories 1. Rye and Products, 2. Oats, and 3. Cereals, Other (Default composition: 68 Popcorn, 89 Buckwheat, 90 Flour, buckwheat, 91 Bran, buckwheat, 92 Quinoa, 94 Fonio, 95 Flour, fonio, 96 Bran, fonio, 97 Triticale, 98 Flour, triticale, 99 Bran, triticale, 101 Canary seed, 103 Grain, mixed, 104 Flour, mixed grain, 105 Bran, mixed grains, 108 Cereals, nes, 111 Flour, cereals, 112 Bran, cereals nes, 113 Cereal preparations, nes)

Pulses (xpul - one MapSPAM category):
- Beans
- Peas
- Pulses, Other and products (Default composition: 181 Broad beans, horse beans, dry, 191 Chick peas, 195 Cow peas, dry, 197 Pigeon peas, 201 Lentils, 203 Bambara beans, 205 Vetches, 210 Lupins, 211 Pulses, nes, 212 Flour, pulses, 213 Bran, pulses)

Notes about other crops:
- sugb (sugar beets) - 0 tonnes (out of 31,096,000 tonnes) in US go to feed. Across the world, 14,405,000/304,420,000 tonnes (5%) go to feed. Sugar is also quite complicated because it combines sugar beet and cane, so probably want to not factor in feed.
- cass (cassava and products) 86,615,000/281,449,000 tonnes (31%) in world go to feed. Maybe include this?
- bana (bananas) 1,011,000/112,069,000 tonnes (0.9%) in world go to feed, so don't include this because it's such a small amount.


MapSPAM to FAO - to compare categories you can look at this df
https://github.com/OHI-Science/food_systems/blob/master/feed/data/MapSPAM_to_FAO_v2.csv

Load libraries
```{r}
library(tidyverse)
library(countrycode)
```

Data Paths
```{r}
raw_path <- "/home/shares/food-systems/social_justice_projects/food_trade_pressures/raw_data/"
FAO_path <- paste0(raw_path, "FAO/")
kastner_path <- paste0(raw_path, "kastner_2022/")
fish_trade_path <- paste0(raw_path, "trade/aquatic_food_trade/")
feed_4a_path <- "/home/shares/food-systems/social_justice_projects/food_trade_pressures/output_data/04a_calc_prop_feed_per_crop_country_FAO_FB_and_CB_2013/"
save_path <- "/home/shares/food-systems/social_justice_projects/food_trade_pressures/output_data/04c_calc_prop_feed_pulses_grains_2017_FAO_FB/"
```


Read in FAO FB data (2010-)
Regions are not included in this df, only countries. To gapfill, will also need to calculate these values for regions.
```{r}
fb <- read_csv(paste0(kastner_path, "CommodityBalances_Crops_2017_E_All_Data_(Normalized)_7_20_22.csv"))

# only 2017 data? - Yes
range(fb$Year)

# clean up data
# converting to tonnes (instead of 1000 tonnes) in case we want to see raw tonnes used for feed later
fb_1 <- fb %>% 
  select(`Area Code (FAO)`, Area, Element, `Item Code`, Item, Unit, Value) %>% 
  filter(Element == "Domestic supply quantity" | Element == "Feed") %>% 
  mutate(value_tonnes = Value*1000) %>% 
  select(-c(Unit, Value))
```


Create function to calculate proportion feed for each category
```{r}
calc_prop_feed <- function(FAO_name, feed_prop_col_name, dsq_col_name, feed_col_name) {
  FAO_name <- sym(FAO_name)
  fb_1 %>% 
  filter(Item == FAO_name) %>% 
  select(-`Item Code`) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes) %>% 
  mutate({{feed_prop_col_name}} := Feed/`Domestic supply quantity`) %>% 
  rename({{dsq_col_name}} := `Domestic supply quantity`) %>% 
  rename({{feed_col_name}} := Feed) %>% 
  select(-Item)
}
```

Run through function with all relevant grains
- Barley and products
- Maize and products
- Sorghum and products
- Millet and products
- Rice and products
- Wheat and products
- Rye and Products
- Oats
- Cereals, Other 
```{r}
barley <- calc_prop_feed("Barley and products", "barl_prop_feed", "barl_dsq", "barl_feed")
maize <- calc_prop_feed("Maize and products", "maiz_prop_feed", "maiz_dsq", "maiz_feed")
sorghum <- calc_prop_feed("Sorghum and products", "sorg_prop_feed", "sorg_dsq", "sorg_feed")
millet <- calc_prop_feed("Millet and products", "mill_prop_feed", "mill_dsq", "mill_feed")
rice <- calc_prop_feed("Rice and products", "rice_prop_feed", "rice_dsq", "rice_feed")
wheat <- calc_prop_feed("Wheat and products", "whea_prop_feed", "whea_dsq", "whea_feed")
rye <- calc_prop_feed("Rye and products", "rye_prop_feed", "rye_dsq", "rye_feed")
oats <- calc_prop_feed("Oats", "oats_prop_feed", "oats_dsq", "oats_feed")
other_cereal <- calc_prop_feed("Cereals, Other", "ocer_prop_feed", "ocer_dsq", "ocer_feed")
#NaN means dsq = 0, so dividing by 0. This could still be feed = 0 / dsq = 0.
# which means that neither 0% nor 100% of dsq really goes to feed. Probably want to gapfill these.
```

Bind grains into one df
```{r}
grains <- barley %>% 
  full_join(maize, by = c("Area Code (FAO)", "Area")) %>% 
  full_join(sorghum, by = c("Area Code (FAO)", "Area")) %>% 
  full_join(millet, by = c("Area Code (FAO)", "Area")) %>% 
  full_join(rice, by = c("Area Code (FAO)", "Area")) %>% 
  full_join(wheat, by = c("Area Code (FAO)", "Area")) %>% 
  full_join(rye, by = c("Area Code (FAO)", "Area")) %>% 
  full_join(oats, by = c("Area Code (FAO)", "Area")) %>% 
  full_join(other_cereal, by = c("Area Code (FAO)", "Area"))

# calculate an average percent feed for grains by summing all feed and dividing by sum of all dsq
grains <- grains %>% 
  rowwise() %>% 
  mutate(total_dsq = sum(barl_dsq, maiz_dsq, sorg_dsq, mill_dsq, rice_dsq, whea_dsq, rye_dsq, oats_dsq, ocer_dsq, na.rm = TRUE)) %>% 
  mutate(total_feed = sum(barl_feed, maiz_feed, sorg_feed, mill_feed, rice_feed, whea_feed, rye_feed, oats_feed, ocer_feed, na.rm = TRUE)) %>% 
  mutate(total_prop_feed = total_feed/total_dsq)

# save grains csv before gapfilling
write_csv(grains, paste0(save_path, "perc_dsq_grains_used_as_feed.csv"))
```

Download FAO FB data for regions for 2017
```{r}
region_fb <- read_csv(paste0(FAO_path, "2017_FAO_data/FAO_Food_Balances_2017_grains_pulses_feed_dl_11_2_22.csv"))

# all 2017? - yes
range(region_fb$Year)

# convert from 1000 tonnes to tonnes
region_fb_1 <- region_fb %>% 
  mutate(value_tonnes = Value*1000) %>% 
  select(-c(`Element Code`, `Item Code (CPC)`, `Year Code`, Year, Unit, Value)) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes) %>% 
  mutate(prop_feed = Feed/`Domestic supply quantity`)

# put in similar format to grains df
region_fb_2 <- region_fb_1 %>% 
  select(`Area Code (M49)`, Area, Item, prop_feed) %>% 
  pivot_wider(names_from = Item, values_from = prop_feed) %>% 
  rename(whea_prop_feed = `Wheat and products`) %>% 
  rename(barl_prop_feed = `Barley and products`) %>% 
  rename(maiz_prop_feed = `Maize and products`) %>% 
  rename(sorg_prop_feed = `Sorghum and products`) %>% 
  rename(rice_prop_feed = `Rice and products`) %>% 
  rename(rye_prop_feed = `Rye and products`) %>% 
  rename(oats_prop_feed = `Oats`) %>% 
  rename(miller_prop_feed = `Millet and products`) %>% 
  rename(ocer_prop_feed = `Cereals, Other`)

# filled in NAs and 0s with continent first, and then world (did this in Excel)
# if the continent has 0 for a crop, then keep it as 0
# if the continent also has NA, then use world value (no continents with NAs in grains df)
# read back in gapfilled csv (only really updated Polynesia, Micronesia, Caribbean, and a couple columns in African regions)
grains_region <- read_csv(paste0(FAO_path, "FAO_regional_grains_perc_feed_gapfilled.csv"))


# calculate aggregate grains per region (sum feed / sum dsq)

# remove pulses
region_grains_wo_pulse <- region_fb_1 %>% 
  filter(!Item == "Beans", !Item == "Pulses, Other and products", !Item == "Peas")

# group by area and sum feed and dsq of grains
region_grains_group <- region_grains_wo_pulse %>% 
  group_by(Area) %>% 
  summarise(sum_dsq = sum(`Domestic supply quantity`, na.rm=TRUE), sum_feed = sum(Feed, na.rm=TRUE))

# check
sum(region_grains_group$sum_dsq)
sum(region_grains_wo_pulse$`Domestic supply quantity`)

# calculate aggregate grain proportion
region_grains_group <- region_grains_group %>% 
  mutate(total_prop_feed = sum_feed/sum_dsq)

# add aggregate grain proportion feed to the rest of the regional individual grain feed df
grains_region <- full_join(grains_region, region_grains_group, by = "Area")
```


Need to add in missing countries that aren't even rows in these data
Combination of FAO country names and additional countries from fisheries trade data.
We won't need all of these countries, but it's good to include them all just in case.
```{r}
# rather than reading in the whole FAO country list, we'll just consider countries included in trade matrices output
# read in FAO country list
tm <- read_csv(paste0(kastner_path, "trade_matrix_by_country_of_origin_2017_import.csv"))

tm_1 <- tm %>% 
  select(Consumer_Country_Code) %>% 
  unique() %>% 
  mutate(country = countrycode(Consumer_Country_Code, "fao", "country.name"))

tm_2 <- tm_1 %>% 
  mutate(country = case_when(Consumer_Country_Code == 41 ~ "China, mainland",
                             Consumer_Country_Code == 299 ~ "Palestine",
                             TRUE ~ country))

# read in fisheries trade data 
fish_tm <- read_csv(paste0(fish_trade_path, "20230420_nceas_consumption.csv"))

# add country names and subset
fish_tm_1 <- fish_tm %>% 
  select(consumer_iso3c) %>% 
  unique() %>% 
  mutate(country = countrycode(consumer_iso3c, 'iso3c', 'country.name')) %>% 
  filter(!is.na(country)) # remove one NA (iso3c = NEI, Netherlands East Indies, not in FAO data either)


# combine dfs to get complete list of countries
countries <- full_join(tm_2, fish_tm_1, by = "country")

# add iso3c
countries_1 <- countries %>% 
  mutate(iso3 = countrycode(country, "country.name", "iso3c")) %>% 
  select(country, iso3)
```


Merge grains df with list of countries from trade matrices to ensure all countries are gapfilled even if there wasn't any FAO FB or CB data for them
```{r}
# add iso3 to grains
grains_a <- grains %>% 
  mutate(iso3 = countrycode(Area, "country.name", "iso3c")) %>% 
  mutate(iso3 = case_when(is.na(iso3) ~ "TUR", TRUE ~ iso3))

# join with country list
grains_b <- grains_a %>% 
  full_join(countries_1, by = "iso3")

# fill in country name where NA
grains_c <- grains_b %>% 
  mutate(Area = case_when(is.na(Area) ~ country, TRUE ~ Area)) %>% 
  select(-country)

# check iso3 codes for duplicates
# remove duplicates: China, China, mainland, Palestinian Territories
# this gets rid of duplicates of China and mainland China
grains_d <- grains_c %>% 
  unique()

# remove Palestinian Territories because we have Palestine
grains_e <- grains_d %>% 
  filter(!Area == "Palestinian Territories")
```


Merge grains df with region by country name (countrycode package)
```{r}
# we need a combination of UN subregion and intermediate codes
# add subregion
grains_e$subregion <- countrycode(grains_e$iso3, origin = 'iso3c', destination = 'un.regionsub.name')

# add intermediate
grains_e$intermed_region <- countrycode(grains_e$iso3, origin = 'iso3c', destination = 'un.regionintermediate.name')

# combine regions
grains_1 <- grains_e %>% 
  mutate(region = case_when(is.na(intermed_region) ~ subregion,
                            TRUE ~ intermed_region))

# compare region categories to what we downloaded from FAO
dplyr::setdiff(grains_region$Area, grains_1$region) # continents are in grains_region but not grains_1
dplyr::setdiff(grains_1$region, grains_region$Area) #nothing in grains_1 that's not in grains_region

# remove subregion and intermediate region categories
grains_2 <- grains_1 %>% 
  select(-c(subregion, intermed_region))

# add continent
grains_2$continent <- countrycode(grains_2$iso3, origin = 'iso3c', destination = 'continent')

# reorder columns so regions is next to country
grains_2 <- grains_2[, c(1:2, 33:35, 3:32)]

# fill in NA for region that didn't match properly (China Taiwan)
grains_2$region[grains_2$`Area Code (FAO)` == 214] <- "Eastern Asia"
```

Gapfill missing or 0 grains data with regional grains feed data
Countries with 0% of a crop going to feed likely just don't have any dsq of that crop.
If they do, then it's likely that at least some of it is going to feed and just not being accounted for.
```{r}
grains_3 <- grains_2

# match regions between 2 dfs
inds <- match(grains_2$region, grains_region$Area)

# fill in NAs and 0s with region level data
grains_3$barl_prop_feed <- ifelse(is.na(grains_3$barl_prop_feed), grains_region$barl_prop_feed[inds], grains_3$barl_prop_feed)
grains_3$barl_prop_feed <- ifelse(grains_3$barl_prop_feed == 0, grains_region$barl_prop_feed[inds], grains_3$barl_prop_feed)

grains_3$maiz_prop_feed <- ifelse(is.na(grains_3$maiz_prop_feed), grains_region$maiz_prop_feed[inds], grains_3$maiz_prop_feed)
grains_3$maiz_prop_feed <- ifelse(grains_3$maiz_prop_feed == 0, grains_region$maiz_prop_feed[inds], grains_3$maiz_prop_feed)

grains_3$sorg_prop_feed <- ifelse(is.na(grains_3$sorg_prop_feed), grains_region$sorg_prop_feed[inds], grains_3$sorg_prop_feed)
grains_3$sorg_prop_feed <- ifelse(grains_3$sorg_prop_feed == 0, grains_region$sorg_prop_feed[inds], grains_3$sorg_prop_feed)

grains_3$mill_prop_feed <- ifelse(is.na(grains_3$mill_prop_feed), grains_region$mill_prop_feed[inds], grains_3$mill_prop_feed)
grains_3$mill_prop_feed <- ifelse(grains_3$mill_prop_feed == 0, grains_region$mill_prop_feed[inds], grains_3$mill_prop_feed)

grains_3$rice_prop_feed <- ifelse(is.na(grains_3$rice_prop_feed), grains_region$rice_prop_feed[inds], grains_3$rice_prop_feed)
grains_3$rice_prop_feed <- ifelse(grains_3$rice_prop_feed == 0, grains_region$rice_prop_feed[inds], grains_3$rice_prop_feed)

grains_3$whea_prop_feed <- ifelse(is.na(grains_3$whea_prop_feed), grains_region$whea_prop_feed[inds], grains_3$whea_prop_feed)
grains_3$whea_prop_feed <- ifelse(grains_3$whea_prop_feed == 0, grains_region$whea_prop_feed[inds], grains_3$whea_prop_feed)

grains_3$rye_prop_feed <- ifelse(is.na(grains_3$rye_prop_feed), grains_region$rye_prop_feed[inds], grains_3$rye_prop_feed)
grains_3$rye_prop_feed <- ifelse(grains_3$rye_prop_feed == 0, grains_region$rye_prop_feed[inds], grains_3$rye_prop_feed)
# the proportion of the dsq going to feed in Eastern Africa is 2.0 because the dsq is only 1,000 tonnes while
# the proportion going to feed is 2,000 tonnes
# convert these 2 values to 1
grains_3$rye_prop_feed <- ifelse(grains_3$rye_prop_feed > 1, 1, grains_3$rye_prop_feed)

grains_3$oats_prop_feed <- ifelse(is.na(grains_3$oats_prop_feed), grains_region$oats_prop_feed[inds], grains_3$oats_prop_feed)
grains_3$oats_prop_feed <- ifelse(grains_3$oats_prop_feed == 0, grains_region$oats_prop_feed[inds], grains_3$oats_prop_feed)

grains_3$ocer_prop_feed <- ifelse(is.na(grains_3$ocer_prop_feed), grains_region$ocer_prop_feed[inds], grains_3$ocer_prop_feed)
grains_3$ocer_prop_feed <- ifelse(grains_3$ocer_prop_feed == 0, grains_region$ocer_prop_feed[inds], grains_3$ocer_prop_feed)
# there are some rows with negative values or infinities for ocer_prop_feed
# convert negative to region value and infinities to 1
grains_3$ocer_prop_feed <- ifelse(grains_3$ocer_prop_feed < 0, grains_region$ocer_prop_feed[inds], grains_3$ocer_prop_feed)
grains_3$ocer_prop_feed <- ifelse(grains_3$ocer_prop_feed > 1, 1, grains_3$ocer_prop_feed)

grains_3$total_prop_feed <- ifelse(is.na(grains_3$total_prop_feed), grains_region$total_prop_feed[inds], grains_3$total_prop_feed)
grains_3$total_prop_feed <- ifelse(grains_3$total_prop_feed == 0, grains_region$total_prop_feed[inds], grains_3$total_prop_feed)

# save gapfilled grains df
write_csv(grains_3, paste0(save_path, "perc_dsq_grains_used_as_feed_gapfilled.csv"))
```



### PULSES

For pulses, need to combine all tonnes and then calculate proportions to make sure it's weighted correctly since we only have one xpul category for MapSPAM.

```{r}
# format beans
beans <- fb_1 %>% 
  filter(Item == "Beans") %>% 
  select(-c(`Item Code`, Item)) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes) %>% 
  rename(beans_dsq = `Domestic supply quantity`,
         beans_feed = Feed)

# format peas
peas <- fb_1 %>% 
  filter(Item == "Peas") %>% 
  select(-c(`Item Code`, Item)) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes) %>% 
  rename(peas_dsq = `Domestic supply quantity`,
         peas_feed = Feed)

# format Pulses, Other and Products
other_pulse <- fb_1 %>% 
  filter(Item == "Pulses, Other and products") %>% 
  select(-c(`Item Code`, Item)) %>% 
  pivot_wider(names_from = Element, values_from = value_tonnes) %>% 
  rename(other_pulse_dsq = `Domestic supply quantity`,
         other_pulse_feed = Feed)

# bind together
pulses <- beans %>% 
  full_join(peas, by = c("Area Code (FAO)", "Area")) %>% 
  full_join(other_pulse, by = c("Area Code (FAO)", "Area"))

# combine dsq and feed for each crop, then calculate proportion feed from total
# have to use rowwise - if we just sum those columns without rowwise and na.rm=TRUE, then if there is an NA
# in any of the columns, we get an NA for the sum
# if pulse_dsq = 0, then prop_feed = NaN, otherwise prop_feed will = 0 if no values (0s or NAs) in pulse_feed
pulses <- pulses %>% 
  rowwise() %>% 
  mutate(pulse_dsq = sum(beans_dsq, peas_dsq, other_pulse_dsq, na.rm=TRUE)) %>% 
  mutate(pulse_feed = sum(beans_feed, peas_feed, other_pulse_feed, na.rm=TRUE)) %>% 
  mutate(prop_feed = pulse_feed/pulse_dsq)

# save pulses df before gapfilling
write_csv(pulses, paste0(save_path, "perc_dsq_pulses_used_as_feed.csv"))
```


Merge pulses df with list of countries from trade matrices to ensure all countries are gapfilled even if there wasn't any FAO FB or CB data for them
```{r}
# add iso3 to pulses
pulses_a <- pulses %>% 
  mutate(iso3 = countrycode(Area, "country.name", "iso3c")) %>% 
  mutate(iso3 = case_when(is.na(iso3) ~ "TUR", 
                          TRUE ~ iso3))

# join with country list
pulses_b <- pulses_a %>% 
  full_join(countries_1, by = "iso3")

# fill in country name where NA
pulses_c <- pulses_b %>% 
  mutate(Area = case_when(is.na(Area) ~ country, 
                          TRUE ~ Area)) %>% 
  select(-country)

# check iso3 codes for duplicates
# remove duplicates: China, China, mainland, Palestinian Territories
# this gets rid of duplicates of China and mainland China
pulses_d <- pulses_c %>% 
  unique()

# remove Palestinian Territories because we have Palestine
pulses_e <- pulses_d %>% 
  filter(!Area == "Palestinian Territories")
```


Merge pulses df with region by country name (countrycode package)
```{r}
# we need a combination of UN subregion and intermediate codes
# add subregion
pulses_e$subregion <- countrycode(pulses_e$iso3, origin = 'iso3c', destination = 'un.regionsub.name')

# add intermediate
pulses_e$intermed_region <- countrycode(pulses_e$iso3, origin = 'iso3c', destination = 'un.regionintermediate.name')

# combine regions
pulses_1 <- pulses_e %>% 
  mutate(region = case_when(is.na(intermed_region) ~ subregion,
                            TRUE ~ intermed_region))

# compare region categories to what we downloaded from FAO
dplyr::setdiff(grains_region$Area, pulses_1$region) # continents are in grains_region but not grains_1
dplyr::setdiff(pulses_1$region, grains_region$Area) #nothing in grains_1 that's not in grains_region

# remove subregion and intermediate region categories
pulses_2 <- pulses_1 %>% 
  select(-c(subregion, intermed_region))

# add continent
pulses_2$continent <- countrycode(pulses_2$iso3, origin = 'iso3c', destination = 'continent')

# reorder columns so regions is next to country
pulses_2 <- pulses_2[, c(1:2, 12:14, 3:11)]

# fill in NA for region that didn't match properly (China Taiwan)
pulses_2$region[pulses_2$`Area Code (FAO)` == 214] <- "Eastern Asia"
```


Calculate percent feed for pulses for each region
Add dsq for beans, peas, and other pulses, do the same with feed, and then divide the aggregate feed by aggregates tonnes.
```{r}
region_beans <- region_fb_1 %>% 
  filter(Item == "Beans") %>%
  select(-c(prop_feed, Item)) %>% 
  rename(beans_dsq = `Domestic supply quantity`,
         beans_feed = Feed)

region_peas <- region_fb_1 %>% 
  filter(Item == "Peas") %>% 
  select(-c(prop_feed, Item)) %>% 
  rename(peas_dsq = `Domestic supply quantity`,
         peas_feed = Feed)

region_other_pulses <- region_fb_1 %>% 
  filter(Item == "Pulses, Other and products") %>%
  select(-c(prop_feed, Item)) %>% 
  rename(other_pulse_dsq = `Domestic supply quantity`,
         other_pulse_feed = Feed)

# combine dfs
region_pulses <- region_beans %>% 
  full_join(region_peas, by = c("Area Code (M49)", "Area")) %>% 
  full_join(region_other_pulses, by = c("Area Code (M49)", "Area"))

# sum dsq and feed and calculate proportion feed
region_pulses <- region_pulses %>% 
  rowwise() %>% 
  mutate(pulse_dsq = sum(beans_dsq, peas_dsq, other_pulse_dsq, na.rm=TRUE)) %>% 
  mutate(pulse_feed = sum(beans_feed, peas_feed, other_pulse_feed, na.rm=TRUE)) %>% 
  mutate(prop_feed = pulse_feed/pulse_dsq)

# set Micronesia to 0 (is NA but should  be 0)
region_pulses <- region_pulses %>% 
  mutate(prop_feed = case_when(Area == "Micronesia" ~ 0, TRUE ~ prop_feed))
```


Gapfill missing or 0 pulses data with regional pulses feed data
Countries with 0% of a crop going to feed likely just don't have any dsq of that crop.
If they do, then it's likely that at least some of it is going to feed and just not being accounted for.
```{r}
pulses_3 <- pulses_2

# match regions between 2 dfs
inds_p <- match(pulses_3$region, region_pulses$Area)

# fill in NAs and 0s with region level data
pulses_3$prop_feed <- ifelse(is.na(pulses_3$prop_feed), region_pulses$prop_feed[inds_p], pulses_3$prop_feed)
pulses_3$prop_feed <- ifelse(pulses_3$prop_feed == 0, region_pulses$prop_feed[inds_p], pulses_3$prop_feed)

# one negative - convert negative to region value
pulses_3$prop_feed <- ifelse(pulses_3$prop_feed < 0, region_pulses$prop_feed[inds_p], pulses_3$prop_feed)

# two values over 1, convert these to 1
pulses_3$prop_feed <- ifelse(pulses_3$prop_feed > 1, 1, pulses_3$prop_feed)

# save gapfilled pulses df
write_csv(pulses_3, paste0(save_path, "perc_dsq_pulses_used_as_feed_gapfilled.csv"))
```
